更新

+++ install_files/install.sh
@@ -1,3 +1,4 @@
 cp /etc/nftclash/install/service.sh /etc/nftclash/service.sh && chmod 755 /etc/nftclash/service.sh
 cp /etc/nftclash/install/nftclashservice /etc/init.d/nftclash && chmod 755 /etc/init.d/nftclash && service nftclash enable
-cp /etc/nftclash/install/version /etc/nftclash/version
+[ ! -f /etc/nftclash/version ] && cp /etc/nftclash/install/version /etc/nftclash/version
+
